pyproject.toml
[tool.poetry]
name = "lms-project"
version = "0.1.0"
description = "LMS system for online learning platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1"
djangorestframework = "^3.15"
pillow = "^10.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3"
drf-spectacular = "^0.27"
stripe = "^10.0"
celery = "^5.3"
redis = "^5.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"